Added TestGlobalFunctions. Split TestMfemHesthaven1D. Fixes.
