+++ test/CMakeLists.txt
@@ -14,14 +14,16 @@ add_executable(maxwell_tests
 "maxwell/TestBilinearIntegrators.cpp" 
 "maxwell/TestMaterial.cpp" 
 "maxwell/TestProbes.cpp" 
-"maxwell/TestSources.cpp")
+"maxwell/TestSources.cpp" 
+"TestGlobalFunctions.cpp" "TestGlobalFunctions.h")
 
 target_link_libraries(maxwell_tests maxwell_solvers GTest::gtest GTest::gtest_main)
 
 add_executable(mfem_hesthaven_tests
 "MFEMHesthavenComparison/TestMfemHesthaven1D.cpp" 
 "MFEMHesthavenComparison/TestMfemHesthaven3D.cpp" 
-"MFEMHesthavenComparison/TestMfemHesthavenFunctions.cpp" "MFEMHesthavenComparison/TestMfemHesthavenFunctions.h")
+"MFEMHesthavenComparison/TestMfemHesthavenFunctions.cpp" "MFEMHesthavenComparison/TestMfemHesthavenFunctions.h" 
+"TestGlobalFunctions.cpp" "TestGlobalFunctions.h" "MFEMHesthavenComparison/TestMfemHesthavenTrace1D.cpp")
 
 target_link_libraries(mfem_hesthaven_tests maxwell_solvers GTest::gtest GTest::gtest_main)
