+++ test/MFEMHesthavenComparison/TestMfemHesthaven3D.cpp
@@ -33,11 +33,13 @@ protected:
 
 };
 
-TEST_F(MFEMHesthaven3D, checkDOperator3DO2)
-{
-	set3DFES(2);
-
-	auto MISCalcOld = 0.5 * buildInverseMassMatrixEigen(fes_) * buildStiffnessMatrixEigen(fes_);
-
-	EXPECT_TRUE(MISCalcOld.isApprox(build3DOneElementDMatrix(), tol_));
-}
+//TEST_F(MFEMHesthaven3D, checkDOperator3DO2)
+//{
+//	set3DFES(2);
+//
+//	auto MISCalcOld = 0.5 * buildInverseMassMatrixEigen(fes_) * buildStiffnessMatrixEigen(fes_);
+//
+//	std::cout << buildStiffnessMatrixEigen(fes_) << std::endl;
+//
+//	EXPECT_TRUE(MISCalcOld.isApprox(build3DOneElementDMatrix(), tol_));
+//}
