+++ test/MFEMHesthavenComparison/TestMfemHesthavenFunctions.h
@@ -13,10 +13,6 @@
 
 
 std::unique_ptr<FiniteElementSpace> buildFiniteElementSpace(const int order);
-Eigen::MatrixXd convertMFEMDenseToEigen(DenseMatrix* mat);
-Eigen::MatrixXd buildMassMatrixEigen(std::unique_ptr<FiniteElementSpace>& fes);
-Eigen::MatrixXd buildInverseMassMatrixEigen(std::unique_ptr<FiniteElementSpace>& fes);
-Eigen::MatrixXd buildStiffnessMatrixEigen(std::unique_ptr<FiniteElementSpace>& fes);
 Eigen::MatrixXd	buildNormalPECFluxOperator1D(std::unique_ptr<FiniteElementSpace>& fes, std::vector<maxwell::Direction> dirVec);
 Eigen::MatrixXd buildExpectedAverageDenseMatrix1D(const int order,const int elements);
 Eigen::MatrixXd buildExpectedJumpDenseMatrix1D(const int order, const int elements);
@@ -24,7 +20,6 @@ Eigen::MatrixXd buildEigenDGTrace1D(std::unique_ptr<FiniteElementSpace>& fes, st
 Eigen::MatrixXd buildEigenMaxwellDGTrace1D(std::unique_ptr<FiniteElementSpace>& fes, std::vector<maxwell::Direction> dir, const double beta);
 std::unique_ptr<BilinearForm> buildBilinearFormWith1DCartesianMesh(const int elements, const int order, std::pair<double, double> ab);
 std::unique_ptr<BilinearForm> buildMaxwellBilinearFormWith1DCartesianMesh(const int elements, const int order, std::vector<maxwell::Direction> dir, const double beta);
-void checkDenseMatrixSubtractIsValueForAllElem(const double val, std::unique_ptr<DenseMatrix> m1, std::unique_ptr<DenseMatrix> m2);
 Eigen::Matrix<double, 27, 27> build3DOneElementDMatrix();
 
 #endif
